create_ticket_for: Autoselect queue, potentially from given KBItem
create_ticket_for now only requires the ticket title and the list of
objects to link to the ticket.

If a 'kbitem' is given, it is linked to the ticket.
If a 'queue' is given in the kwargs the ticket is filed to this queue.
Otherwise if a 'kbitem' is given, the queue of its KBCategory is used.
As a fallback the first queue found is used.